Make wikimetrics less Linux-centric, more UNIX-agnostic
Python could be installed on different place on different OSes.

For example, it's located at /usr/local/bin/python on FreeBSD.

On a Solaris machine, it could be at an arbitrary path, like
/opt/ts/python/<version>/bin/python

The best practice is so to find python in the enviroment prompt,
instead to hardcore a path. We then call /usr/bin/env (which is
a standard executable in every POSIX OS, so always there).

Bug: 57494
Change-Id: Icd86863bcc54f5ce49c3bb07df784a38b3ad0f13