Fix user name display in CSV files
In order to test this change, we added a new database.  The corresponding
puppet configuration was changed to add this in vagrant and production.
The new database is called wiki2_testing and can be referenced from
tests/fixtures.py as second_mediawiki_project.  The new tests make sure
that cross-project cohorts are safe.  They also make sure that wiki
users which share a name or id within or accross cohorts do not break
the new way in which we're handling report results.

Related Puppet changes:
    https://gerrit.wikimedia.org/r/#/c/129542/

Bug: 64026
Change-Id: I1d3915df9ed67c9bf3f1dc6fa901cbdd5803629b