Move wikimetrics database models to models/storage
This is step 1 of implementing cohorts whose memberships are entire
wiki projects.  That feature will require a class hierarchy of logical
Cohort objects to exist, and so we have to first make it easy to tell
the difference between the new objects and the existing storage objects
called Cohort.  Since doing this *just* for Cohort would make for a very
inconsistent codebase, I just moved all the wikimetrics database table
mapping objects to models/storage and appended "Store" to their names.

Change-Id: I177f3f91b13d0128bc310bab3202208f2d84f719