Make tests in wikimetrics more deterministic.
We are running celery on sync mode (CELERY_ALWAYS_EAGER)
for the tests that is possible to do so, which are the majority.
The notable exception are the tests for the report controller.

We are not starting the scheduler with the tests
any more as it is not needed. We have removed the
nonDeterministic and sleeps from scheduler tests.

Change-Id: I14eb5e902e92067e56497d5d68f59e132967a2aa