Add rolling active editor
http://meta.wikimedia.org/wiki/Research:Rolling_active_editor

Timeseries support for this is trickier than trivial so I think we
should wait.  Basically, timeseries support needs to handle two distinct
cases: "count" metrics and "boolean" metrics.  These work differently
and if we build support for timeseries "boolean" metrics, we should do
it generally and not just hack it in.  The tests for timeseries rolling
monthly active editors are saved in an old patcshet (PS5) and referenced
from a comment.

Bug: 67458
Change-Id: I9ef4d9cebb8575efddd7f1704a6180b1c4f9afeb