Make test script not complain if no *.pyc files exist
If no *.pyc files exist in the tests directory, “rm” was called
without arguments. “rm” in turn complained about missing arguments.
To avoid this, we teach xargs to not run “rm” if there are no
arguments.

The “--no-run-if-empty” is a GNU extension, but it's really
handy. Working around the “empty arguments” issue without this option
is more twisted than necessary.

Change-Id: Ieb0d18bf1dbe14cb5357a21b816eccf37e6246c9