Bring connection pool tests into main test suite
The datadase connection pool tests do not require manual setup, they
can run on a plain setup, so bring them from the manual into the main
test suite.

Since one of the two tests is slow, we add a slow group and exclude it
by default. That way, it does not get in the way when developing, but
we can check it automatically before deployment.

Change-Id: Ibbbec4e432a5dba65b2a85cc0a3aa4a622a56d1e