Allow to run nosetests without coverage report
While coverage reports are usefull, they also easily fill up a whole
screen and burry relevant information in scrollback, which gets in the
way when developing.
By setting the “with-coverage” in setup.cfg, we forced those coverage
reports even when running nosetests directly and coverage reports
could not be turned off.

Hence, we remove this forced setting from setup.cfg, and thereby allow
running nosetests with or without coverage report at user's choice.

Since the option is duplicated to tox.ini, we drop it from there too.

scripts/test is unaffected by this change and continues to show
coverage report, as it passes the --with-coverage option directly.

Change-Id: I0d53bbe96bf465df77f05675d1bf1e3d85ac3a02