Fix slow Rolling Active Editor metric
There was a performance mistake in the previous version of the query.
The query was filtering on users only after collecting data from the
revision and archive tables.  Filtering on users closer to the
subqueries against revision and archive tables speeds up execution a lot
because there of the indices on rev_user and ar_user.

Bug: 68596
Change-Id: Id992ee2030d6257eb681dff2cf4246750097f9af