Reschedule recurring reports to 03:00
We ran the recurring reports an 00:00 which meant that even a lag of a
few minutes might falsify the reports we run. Hence, we now switch to
running them at 03:00. This rescheduling does not solve the lag issue,
but it gives us a at least a bit of safety buffer.

Bug: 68507
Change-Id: I186587296e6e8e16a35e1e9ac76f0d9cfa9458c9