Stop scheduling new recurrent runs if databases lag
We scheduled recurrent runs regardless of the databases' replication
lag. But if there is a replication lag, this would falsify the results
of the runs. Hence, we now don't schedule new recurrent runs if any of
the databases in REPLICATION_LAG_MW_PROJECTS show replication lag.
Those recurrent runs will automatically get scheduled on the next
scheduler run once the database replication caught up.

The credentials that we use to connect to the databases lack grants to
use MariaDB's built in measures for lag detection. So, we check for
the most recent edit instead. If this edit is older than
REPLICATION_LAG_THRESHOLD hours, we consider the database lagged,
otherwise, we assume the replication lag is fine.

This change does not address replication lag issues for non-recurrent
runs.

Bug: 68507
Change-Id: I31974426ed55f4719b764885e550d98c8c525e4c