Disable pooling for mediawiki dbs
Connecting to the mediawiki dbs with database-specific URIs means
pooling works at the database level.  This used to be necessary but now
labsdb has a single-box setup which would enable us to use the same URI
for all dbs, and just issue `use enwiki_p;` as part of our query for
example.  This, however, will be left for later.  For now, disabling
pooling eliminates problems with starving the allowed connections to our
labsdb user.

Bug: 68833
Change-Id: I224aedd4893f1dc109907dc37cad1bf72afd7975