Delete PENDING reports and increase Timeout
If recurrent report children are in the PENDING state for more than a
few seconds, it means they should be cleaned up.  This is because delay
is called on their task and the first step there is to change the status
to STARTED.  Therefore, we're deleting any PENDING children before we
figure out which days are left to run.

Also, related, the timeout on the daily scheduled task is increased to
three times the default.  This is to accomodate the fact that the
scheduler task won't execute as quickly since celery is swamped by
handling all the new workers that the scheduler is putting to work.

Bug: 70663
Change-Id: I9e8ef523833c9395ef250e12757c449e2fb32b6b