Default to all namespaces for edits and pages
Changes the default metric parameters for PagesCreated and
NamespaceEdits to look at all namespaces.  Deployment of this will be
combined with a thorough announcement to the wikimetrics list, as the
change will cause most numbers to increase unless people manually enter
the previous default, of only looking at namespace zero.

Bug: 72114
Change-Id: Ia7192f8c6cbafb2196084281bbc8dd863c16eecd