Do not store reports that are not going to be used
When creating a report from the gui, the system internally creates
various reports and stores them in the database. One is the actual
final report, and the others are secondary reports used by the latter.
In some cases, i.e. when using centralauth expansion, the system may
create lots of secondary reports in the database (>500), and they
are not being used today.

This fix avoids storing these secondary reports in the database.
For that, a parameter has been added to the Report class initializer.
This parameter tells if the report should be stored in the database.
For primary reports, the initialized will be called with the parameter
set to True. For secondary-, the parameter will default to False.

Bug: 72635
Change-Id: Ic47542f940bb40462263a8ac21cc5499537ba16d