Add default tags and default project to cohorts view
This change adds a script that inserts all the predefined tags
into the tag table in the wikimetrics database. The input file
for the script is stored in the bug track. See link/code below.

This change also adds the default project to the cohort detail
view, together with a notification whenever the cohort has been
uploaded with centralauth expansion.

Bug: 72746
Change-Id: I2ee332698e6f223370e7def1b8922f4b41359648