Add cohort membership page
Adds a cohort membership page, where the cohort users
are listed, can be filtered (by name and by validity)
and can be deleted. The new page is accessible from
the cohort view, through a button named 'View Members'.

The users in the membership page are grouped by username.
Each record features: username, projects, validity and
a button to remove the user from the cohort.
The page has no pagination, it shows the 20 first users
that accomplish the filter conditions.

Note that a username can have an account in various wikis!
Sometimes the user may be valid, and sometimes invalid.
And even sometimes, the user may be valid in some of its
wikis and invalid in others.

In this last case, the row shows another button named
'Remove Invalid' that deletes only the part of the user
that is invalid.

This fix also removes the invalid wikiusers page,
because the membership page includes all of its
functionalities. Its tests were removed also.

Bug: T75350
Change-Id: I588e03612cbcef25117daad98ce90576d9b9fffb