Fix up tox setup and setup.py parse_requirements()
Seems some version of pip now require a Pip.Download session to be
setup. I have no idea what it is but that can be ignored by setting
session=False.

For tox:

Since we have skipsdist = True, we need to set usedevelop = True to have
the console scripts installed in the venv ./bin/ directory.  Else the
wikimetrics CLI can't be found.

Finally, append {posargs} to the nosetests command, so we can pass it
additional parameters. Example:

 tox -epy27 -- --help

Will shows the nosetest help :-)

Change-Id: I0f7f9a25b7b9e36d9f93301b9de1a6cc575cc563