Add all timezones to the timezone picker
It was suggested in the Task that we add 24 time zones and put up a
message that helps people find information about timezones.  I instead
added all 40 valid time zones as it doesn't make the drop down any
harder to work with.  I also added a generic help link to the Wikipedia
time zone article that explains UTC offsets.

Bug: T121167
Change-Id: Ie82fa91db4d74f5a9f6a0667d90e29efe1fb6224