Clear connections between report executions
The max connections per user in labsdb databases is 10 for the
Wikimetrics user. This prevents some centralauth-expanded cohorts
to be executed. Because Wikimetrics tries to connect to each
database with a different connection to the same host.

This patch clears the connections after each report node has finished
executing. This allows report on 'wide' centralauth-expanded cohorts
to be run.

This is not the ideal behavior of Wikimetrics database connections.
However, given that the report tree is already executed sequentially,
the connections do not need to last for longer.

Bug: T173585
Change-Id: I69d38cab9103e2a69f070471a0af662829a66eba